# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auth",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://api.marketfitcolombia.com"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "summary": "Registrar usuarios",
                "operationId": "ad77fe138bcb384ad65d167d5c8d1cc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registra usuario y retorna info y token."
                    },
                    "400": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "summary": "Iniciar sesión",
                "operationId": "c452e49acfec9a0c37276b0c1d13bf9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicia sesión y retorna token."
                    },
                    "400": {
                        "description": "Credenciales inválidas."
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "summary": "Cerrar sesión",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cerrar sesión exitosamente."
                    },
                    "400": {
                        "description": "Error al cerrar sesión."
                    }
                }
            }
        },
        "/api/v1/user": {
            "post": {
                "summary": "Obtener información del usuario",
                "operationId": "614a24dcd8f3eee2a2cae1682ea8ac72",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información del usuario obtenida."
                    },
                    "401": {
                        "description": "Token inválido o expirado."
                    }
                }
            }
        },
        "/api/v1/user/update": {
            "post": {
                "summary": "Actualizar información del usuario",
                "operationId": "0f4fd3fbbeb89103f3d0441e9f970897",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente."
                    },
                    "401": {
                        "description": "Token inválido o expirado."
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "summary": "Obtener todas las categorías",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "responses": {
                    "200": {
                        "description": "Retorna todas las categorías."
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva categoría",
                "operationId": "dd76b8d73b7ea8b4951f03d7c0904c92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría creada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la categoría."
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "summary": "Obtener una categoría por su ID",
                "operationId": "9c4934d1b68a6d4440ec72cfc8ae7074",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "La categoría no existe."
                    }
                }
            },
            "put": {
                "summary": "Actualizar una categoría por su ID",
                "operationId": "988e8695bc991d7f8e40131db5ba7a76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "La categoría no existe."
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una categoría por su ID",
                "operationId": "37196d259228d2274dd9dbef8b00e547",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "La categoría no existe."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Categoria": {
                "title": "Categoria",
                "properties": {
                    "nombre": {
                        "description": "Nombre de la categoría",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción de la categoría",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}